chore: Refactor email sending to skip sending emails in test mode
