+++ occupi-backend/tests/middleware_test.go
@@ -4,7 +4,9 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"strings"
+	"sync"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
@@ -14,6 +16,7 @@ import (
 	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/authenticator"
 	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/constants"
 	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/database"
+	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/middleware"
 	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/router"
 	// "github.com/stretchr/testify/mock"
 )
@@ -281,4 +284,150 @@ func TestAccessUnprotectedRouteWithSessionInvalidToken(t *testing.T) {
 		"{\"data\":null,\"message\":\"pong -> I am alive and kicking and you are not auth'd, only non-auth'd users can access this endpoint\",\"status\":200}",
 		strings.ReplaceAll(w1.Body.String(), "-\\u003e", "->"),
 	)
+}
+
+func TestRateLimit(t *testing.T) {
+	// connect to the database
+	db := database.ConnectToDatabase(constants.AdminDBAccessOption)
+
+	// set gin run mode
+	gin.SetMode(configs.GetGinRunMode())
+
+	// Create a Gin router
+	ginRouter := gin.Default()
+
+	// adding rate limiting middleware
+	middleware.AttachRateLimitMiddleware(ginRouter)
+
+	// Register routes
+	router.OccupiRouter(ginRouter, db)
+
+	server := httptest.NewServer(ginRouter)
+	defer server.Close()
+
+	var wg sync.WaitGroup
+	numRequests := 10
+	responseCodes := make([]int, numRequests)
+
+	for i := 0; i < numRequests; i++ {
+		wg.Add(1)
+		go func(index int) {
+			defer wg.Done()
+			resp, err := http.Get(server.URL + "/ping")
+			if err != nil {
+				t.Errorf("Request %d failed: %v", index, err)
+				return
+			}
+			defer resp.Body.Close()
+			responseCodes[index] = resp.StatusCode
+		}(i)
+		time.Sleep(100 * time.Millisecond) // Slight delay to spread out the requests
+	}
+
+	wg.Wait()
+
+	rateLimitedCount := 0
+	for _, code := range responseCodes {
+		if code == http.StatusTooManyRequests {
+			rateLimitedCount++
+		}
+	}
+
+	assert.Greater(t, rateLimitedCount, 0, "There should be some requests that are rate limited")
+	assert.LessOrEqual(t, rateLimitedCount, numRequests-5, "There should be at least 5 requests that are not rate limited")
+}
+
+func TestRateLimitWithMultipleIPs(t *testing.T) {
+	// connect to the database
+	db := database.ConnectToDatabase(constants.AdminDBAccessOption)
+
+	// set gin run mode
+	gin.SetMode(configs.GetGinRunMode())
+
+	// Create a Gin router
+	ginRouter := gin.Default()
+
+	// adding rate limiting middleware
+	middleware.AttachRateLimitMiddleware(ginRouter)
+
+	// Register routes
+	router.OccupiRouter(ginRouter, db)
+
+	server := httptest.NewServer(ginRouter)
+	defer server.Close()
+
+	var wg sync.WaitGroup
+	numRequests := 10
+	ip1 := "192.168.1.1"
+	ip2 := "192.168.1.2"
+	responseCodesIP1 := make([]int, numRequests)
+	responseCodesIP2 := make([]int, numRequests-5)
+
+	// Send requests from the first IP address
+	for i := 0; i < numRequests; i++ {
+		wg.Add(1)
+		go func(index int) {
+			defer wg.Done()
+			client := &http.Client{}
+			req, err := http.NewRequest("GET", server.URL+"/ping", nil)
+			if err != nil {
+				t.Errorf("Failed to create request: %v", err)
+				return
+			}
+			req.Header.Set("X-Forwarded-For", ip1)
+			resp, err := client.Do(req)
+			if err != nil {
+				t.Errorf("Request failed: %v", err)
+				return
+			}
+			defer resp.Body.Close()
+			responseCodesIP1[index] = resp.StatusCode
+		}(i)
+		time.Sleep(10 * time.Millisecond) // Slight delay to spread out the requests
+	}
+
+	// Send requests from the second IP address
+	for i := 0; i < numRequests-5; i++ {
+		wg.Add(1)
+		go func(index int) {
+			defer wg.Done()
+			client := &http.Client{}
+			req, err := http.NewRequest("GET", server.URL+"/ping", nil)
+			if err != nil {
+				t.Errorf("Failed to create request: %v", err)
+				return
+			}
+			req.Header.Set("X-Forwarded-For", ip2)
+			resp, err := client.Do(req)
+			if err != nil {
+				t.Errorf("Request failed: %v", err)
+				return
+			}
+			defer resp.Body.Close()
+			responseCodesIP2[index] = resp.StatusCode
+		}(i)
+		time.Sleep(10 * time.Millisecond) // Slight delay to spread out the requests
+	}
+
+	wg.Wait()
+
+	rateLimitedCountIP1 := 0
+	rateLimitedCountIP2 := 0
+	for _, code := range responseCodesIP1 {
+		if code == http.StatusTooManyRequests {
+			rateLimitedCountIP1++
+		}
+	}
+	for _, code := range responseCodesIP2 {
+		if code == http.StatusTooManyRequests {
+			rateLimitedCountIP2++
+		}
+	}
+
+	// Assertions for IP1
+	assert.Greater(t, rateLimitedCountIP1, 0, "There should be some requests from IP1 that are rate limited")
+	assert.LessOrEqual(t, rateLimitedCountIP1, numRequests-5, "There should be at least 5 requests from IP1 that are not rate limited")
+
+	// Assertions for IP2
+	assert.Equal(t, rateLimitedCountIP2, 0, "There should be no requests from IP2 that are rate limited")
 }
